Add price to events in double precision format

+++ include/event/FundsCredited.h
@@ -6,12 +6,12 @@ class FundsCredited: public IDomainEvent {
 public:
     static constexpr std::string EVENT_TYPE = "FundsCredited";
 
-    FundsCredited(std::string event_id, int user_id, int quantity);
+    FundsCredited(std::string event_id, int user_id, double quantity);
     explicit FundsCredited(const std::string &json_string);
 
     std::string toJson() const override;
 
 private:
     int user_id_;
-    int quantity_;
+    double quantity_;
 };
